New way of defining configurations
Added a way of setting up specific test values for different groups. The
idea is to store them in a Scripting.Dictionary (key-value pair, like a
hash map), which makes is possible to label a group of values for a
certain group. Should make things easier to handle and keep track of as
the list of test cases grows.